Added extra parameters for dominant_label.
This extra parameter in WeakForm and CanonicalEquation can be used to
overwrite the standard procedure to chose the highest present temporal
derivative to be the dominant variable of an equation.

Nevertheless, be aware that the names you assign to WeakFormulations
have to be unique and will cause mayhem if assigned twice.
(These names act as keys for several dictionaries whose entries will be overridden
by when passing duplicate names)

Long story short, this should fix the introduced TestCase.