Fix issue: "Insufficient number of roots detected"
see commit aa2b38c

Actually the issue is still present try for example
l = 1e6 or l = 1e-6. But for the test example the range
1e-3 < l < 1e3 is ok.

If someone needs this feature for a very large/small
domain, he must take care of the correct scaling of
the tolerances and some floating point errors
(another issue).

I dont have the time for this at the moment.