Some thoughts to stacked base
Since stacked bases have always a different meaning then instances
of Base, for which the whole toolbox/transformation-routine is designed,
their only purpose should be to select the correct portion of weights
with the help of their transformation hint.

In order to prevent misuse I think stacked bases
* should not be scalable
* should be incompatible to everything
* and just for the case that somebody ask,
  the function space should be unique.

All the other stuff can be done with the Base class.