Got the example running again
As discussed, the `masking` of values via products with `index`
test-functions was not really intended.
For the best solution, the created 2d data fields should be checked in
`evaluate_placeholder` for this occurrence and summed up over the
corresponding dimension if all other cols (or rows) are zero (masked).
This saves the extra effort in `compute_product_of_scalars`.
However, due to some cases where the full dimensionality is needed in
later processing steps, this can't be done there.

I tweaked the introduced extra checks to go along with the design idea.
However, the integration of temporal derivatives into the bases
themselves (which I really like) leads to a rather big boilerplate code
overhead. But this is not something to be discussed in this context.